+++ R/Gauss_test_power.R
@@ -1,5 +1,5 @@
 
-#' Computation of power and sufficient sample size for the one-sided t-test
+#' Computation of power and sufficient sample size for the one-sided Gauss test
 #'
 #' Let \mjseqn{X_1, \dots, X_n} be \mjseqn{n} i.i.d. variables
 #' with mean \mjseqn{\mu}, variance \mjseqn{\sigma^2}.
@@ -8,8 +8,10 @@
 #' Let \mjseqn{\eta := (\mu - \mu_0) / \sigma} be the effect size,
 #' i.e. the distance between the null and the alternative hypotheses,
 #' measured in terms of standard deviations.
+#'
 #' There is a relation between the sample size \code{n}, the effect size \code{eta}
-#' and the power \code{beta}.
+#' and the power \code{beta}. This function takes as an input two of these quantities
+#' and return the third one.
 #'
 #' \loadmathjax
 #' @import mathjaxr
@@ -32,6 +34,8 @@
 #' \mjseqn{kappa :=} sup of modulus of \mjseqn{f_{X_n / \sigma_n}(t)}
 #' over all \mjseqn{t} such that \mjseqn{|t| \geq 2 t_1^* \pi / K3tilde}.
 #'
+#' @return The computed value of either the sufficient sample size \code{n},
+#' or the minimum effect size \code{eta}, or the power \code{beta}.
 #'
 #' @references
 #' Derumigny A., Girard L., and Guyonvarch Y. (2021).
@@ -41,21 +45,21 @@
 #' @examples
 #'
 #' # Sufficient sample size to detect an effect of 0.5 standard deviation with probability 80%
-#' t_test_powerAnalysis(eta = 0.5, beta = 0.8)
+#' Gauss_test_powerAnalysis(eta = 0.5, beta = 0.8)
 #' # We can detect an effect of 0.5 standard deviations with probability 80% for n >= 548
 #'
 #' # Power of an experiment to detect an effect of 0.5 with a sample size of n = 800
-#' t_test_powerAnalysis(eta = 0.5, n = 800)
+#' Gauss_test_powerAnalysis(eta = 0.5, n = 800)
 #' # We can detect an effect of 0.5 standard deviations with probability 85.1% for n = 800
 #'
 #' # Smallest effect size that can be detected with a probability of 80% for a sample size of n = 800
-#' t_test_powerAnalysis(n = 800, beta = 0.8)
+#' Gauss_test_powerAnalysis(n = 800, beta = 0.8)
 #' # We can detect an effect of 0.114 standard deviations with probability 80% for n = 800
 #'
 #'
 #' @export
 #'
-t_test_powerAnalysis = function(eta = NULL, n = NULL, beta = NULL,
+Gauss_test_powerAnalysis = function(eta = NULL, n = NULL, beta = NULL,
                                 alpha = 0.05, K4 = 9, kappa = 0.99){
 
   if (is.null(eta) + is.null(n) + is.null(beta) != 1){
@@ -64,24 +68,29 @@ t_test_powerAnalysis = function(eta = NULL, n = NULL, beta = NULL,
   }
 
   if (is.null(beta)){
-    return (t.test.power(eta = eta, n = n,
-                         alpha = alpha, K4 = K4, kappa = kappa))
+    result = Gauss_test_power(eta = eta, n = n,
+                              alpha = alpha, K4 = K4, kappa = kappa)
+    names(result) <- "beta"
+
   } else if (is.null(n)){
 
-    result = stats::uniroot(
-      f = function(n){t.test.power(eta = eta, n = n,
+    res = stats::uniroot(
+      f = function(n){Gauss_test_power(eta = eta, n = n,
                                    alpha = alpha, K4 = K4, kappa = kappa) - beta},
       interval = c(1, 10^10))
 
-    return (ceiling(result$root))
+    result = ceiling(res$root)
+    names(result) <- "n"
+
   } else {
 
-    result = stats::uniroot(
-      f = function(eta){t.test.power(eta = eta, n = n,
+    res = stats::uniroot(
+      f = function(eta){Gauss_test_power(eta = eta, n = n,
                                      alpha = alpha, K4 = K4, kappa = kappa) - beta},
       interval = c(0, 1))
 
-    return (result$root)
+    result = res[["root"]]
+    names(result) <- "eta"
   }
 
   return (result)
@@ -89,7 +98,7 @@ t_test_powerAnalysis = function(eta = NULL, n = NULL, beta = NULL,
 
 
 # compute beta from eta and n
-t.test.power = function(eta, n, alpha, K4, kappa){
+Gauss_test_power = function(eta, n, alpha, K4, kappa){
 
   delta_n = BoundEdgeworth::Bound_BE(
     setup = list(continuity = TRUE, iid = TRUE, no_skewness = FALSE),